VAGOV-000: Clean up links from menu source - remove localhost and va.gov
