+++ docroot/modules/custom/va_gov_migrate/src/Plugin/migrate/source/VaBenefitsMenu.php
@@ -3,7 +3,6 @@
 namespace Drupal\va_gov_migrate\Plugin\migrate\source;
 
 use Drupal\migrate\MigrateException;
-use Drupal\migrate\Plugin\migrate\source\SourcePluginBase;
 use Drupal\migrate\Plugin\MigrationInterface;
 
 /**
@@ -13,7 +12,7 @@ use Drupal\migrate\Plugin\MigrationInterface;
  *   id = "va_benefits_menu_source"
  * )
  */
-class VaBenefitsMenu extends SourcePluginBase {
+class VaBenefitsMenu extends VaMenuBase {
   protected $sections;
 
   /**
@@ -74,198 +73,24 @@ class VaBenefitsMenu extends SourcePluginBase {
       }
     }
 
-    $this->setIds($menus);
-    $this->setWeights($menus);
-    $flat_menu = $this->flattenMenu($menus);
-
-    return new \ArrayIterator($flat_menu);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getIds() {
-    $ids['id']['type'] = 'string';
-    return $ids;
+    return new \ArrayIterator($this->process($menus));
   }
 
   /**
    * {@inheritdoc}
    */
   public function fields() {
-    $fields['href'] = 'Link Path';
-    $fields['title'] = 'Link Title';
-    $fields['external'] = 'External Path';
-    $fields['weight'] = 'Weight';
-    $fields['parent_id'] = 'Parent Id';
+    $fields['href'] = parent::fields();
     $fields['menu'] = 'Menu machine name';
 
     return $fields;
   }
 
   /**
-   * Add weights to menu items based on their order in the array.
-   *
-   * @param array $menu_tree
-   *   The array of menu items.
+   * {@inheritdoc}
    */
-  public static function setWeights(array &$menu_tree) {
-    $relative_weight = 0;
-    foreach ($menu_tree as &$item) {
-      $item['weight'] = -50 + $relative_weight;
-      $relative_weight++;
-      if (!empty($item['items'])) {
-        self::setWeights($item['items']);
-      }
-    }
-  }
-
-  /**
-   * Create a unique id for each menu item.
-   *
-   * The ids are used in post processing to set parent items.
-   *
-   * @param array $menu_tree
-   *   The menu tree to set ids on.
-   * @param string $parent_id
-   *   The id of the current menu tree's parent.
-   */
-  public static function setIds(array &$menu_tree, $parent_id = '') {
-    foreach ($menu_tree as $index => &$item) {
-      $item['id'] = $parent_id . '-' . $index;
-      if (!empty($item['items'])) {
-        self::setIds($item['items'], $item['id']);
-      }
-    }
-  }
-
-  /**
-   * Transform a menu tree into a flat menu with parent href set for children.
-   *
-   * Also clean up the hrefs and set 'external'.
-   *
-   * @param array $menu_tree
-   *   The tree to transform.
-   * @param string $parent_id
-   *   The parent id, if any.
-   * @param string $parent_menu
-   *   The parent menu name (will be filled in during recursion from top item).
-   *
-   * @return array
-   *   A one-dimensional array of menu items.
-   */
-  public static function flattenMenu(array $menu_tree, $parent_id = '', $parent_menu = '') {
-    $flat_menu = [];
-    foreach ($menu_tree as $index => $item) {
-      if (empty($item['href'])) {
-        $item['href'] = 'route:<nolink>';
-        $item['external'] = 0;
-      }
-      elseif (parse_url($item['href'], PHP_URL_SCHEME)) {
-        $item['href'] = str_replace('http://localhost:3001', '', $item['href']);
-        $item['external'] = 1;
-      }
-      else {
-        $item['href'] = rtrim($item['href'], '/');
-        $item['external'] = 0;
-      }
-
-      if (!empty($parent_menu)) {
-        $item['menu'] = $parent_menu;
-      }
-      $item['parent_id'] = $parent_id;
-      $flat_menu[] = $item;
-
-      if (!empty($item['items'])) {
-        $new_parent = '';
-        if (!empty($item['menu'])) {
-          $new_parent = $item['menu'];
-        }
-        $flat_menu = array_merge($flat_menu, self::flattenMenu($item['items'], $item['id'], $new_parent));
-      }
-      unset($item['items']);
-
-    }
-
-    return $flat_menu;
-  }
-
-  /**
-   * Takes two page menus and merges them into one.
-   *
-   * Assumes that no two menu items have the same title & link.
-   *
-   * @param array $menu
-   *   Array of menu trees.
-   * @param array $menu2
-   *   Another array of menu trees.
-   *
-   * @return array
-   *   The consolidated menu tree.
-   *
-   * @throws \Drupal\migrate\MigrateException
-   */
-  protected function mergeMenus(array $menu, array $menu2) {
-    $merge_menu = $menu2;
-
-    foreach ($menu as $menu_item) {
-      $found = FALSE;
-      foreach ($merge_menu as &$merge_item) {
-        if ($menu_item['title'] == $merge_item['title'] &&
-          ((empty($menu_item['href']) && empty($merge_item['href'])) ||
-            $menu_item['href'] == $merge_item['href'])) {
-          if (!empty($menu_item['items'])) {
-            if (empty($merge_item['items'])) {
-              $merge_item['items'] = $menu_item['items'];
-            }
-            else {
-              $merge_item['items'] = $this->mergeMenus($menu_item['items'], $merge_item['items']);
-            }
-          }
-          $found = TRUE;
-          break;
-        }
-      }
-      if (!$found) {
-        $merge_menu[] = $menu_item;
-      }
-    }
-
-    return $merge_menu;
-  }
-
-  /**
-   * Find the parent item for submenus and merge them in.
-   *
-   * @param array $page
-   *   The page structure containing the submenu.
-   * @param array $menu
-   *   The full menu to merge the submenus into.
-   *
-   * @return bool
-   *   True if a parent item was found.
-   *
-   * @throws \Drupal\migrate\MigrateException
-   */
-  protected function findMergeMenus(array $page, array &$menu) {
-    $page_link = $page['backLink']['href'];
-    foreach ($menu as &$menu_item) {
-      if (!empty($menu_item['href']) && $menu_item['href'] == $page_link) {
-        if (empty($menu_item['items'])) {
-          $menu_item['items'] = $page['items'];
-        }
-        else {
-          $menu_item['items'] = $this->mergeMenus($page['items'], $menu_item['items']);
-        }
-        return TRUE;
-      }
-      if (!empty($menu_item['items'])) {
-        if ($this->findMergeMenus($page, $menu_item['items'])) {
-          return TRUE;
-        }
-      }
-    }
-    return FALSE;
+  protected function sanitizeDomain($href) {
+    return str_replace('http://localhost:3001', '', $href);
   }
 
 }
