+++ docroot/modules/custom/va_gov_migrate/src/Plugin/migrate/source/VaMainMenu.php
@@ -2,8 +2,6 @@
 
 namespace Drupal\va_gov_migrate\Plugin\migrate\source;
 
-use Drupal\migrate\Plugin\migrate\source\SourcePluginBase;
-
 /**
  * Source to read from sidebar.json.
  *
@@ -11,7 +9,7 @@ use Drupal\migrate\Plugin\migrate\source\SourcePluginBase;
  *   id = "va_main_menu_source"
  * )
  */
-class VaMainMenu extends SourcePluginBase {
+class VaMainMenu extends VaMenuBase {
 
   /**
    * Return a string representing the source file path.
@@ -46,11 +44,7 @@ class VaMainMenu extends SourcePluginBase {
       $menus[] = $main_section;
     }
 
-    VaBenefitsMenu::setIds($menus);
-    VaBenefitsMenu::setWeights($menus);
-    $flat_menu = VaBenefitsMenu::flattenMenu($menus);
-
-    return new \ArrayIterator($flat_menu);
+    return new \ArrayIterator($this->process($menus));
   }
 
   /**
@@ -64,14 +58,8 @@ class VaMainMenu extends SourcePluginBase {
   /**
    * {@inheritdoc}
    */
-  public function fields() {
-    $fields['href'] = 'Link Path';
-    $fields['title'] = 'Link Title';
-    $fields['external'] = 'External Path';
-    $fields['weight'] = 'Weight';
-    $fields['parent_id'] = 'Parent Id';
-
-    return $fields;
+  protected function sanitizeDomain($href) {
+    return str_replace('https://www.va.gov', '', $href);
   }
 
   /**
